Save capability registers in the trapframe.
- Save capability registers instead of GPRs in the trapframe.
- Construct a csp from the kernel ddc and sp during a return to
  userland that can be used while ddc is set to the user ddc.
  Save this in sscratchc instead of using sscratch.  Entering the
  kernel uses csp from sscratchc to save state before switching to
  the kernel ddc.
- Run exception handler in capmode and switch to non-capmode
  before entering C.  Likewise, switch to capmode before restoring
  user ddc when returning to userland.  This requires moving the
  init of the kernel gp later until after we are in nocapmode.
- Since it is tied heavily to struct trapframe, update mcontext_t as
  well.  The capability version only stores capabilities and no
  GPRs.  The legacy 64-bit version stores ddc and sepcc in the spare
  room, and stores the offset of sepcc in the sepc GPR.  For
  set_mcontext it uses sepc as an offset relative to the saved sepcc
  to construct sepcc.
- Update pcb to hold capability registers instead of GPRs.
- Implement fill_capregs().
- Partial updates to sendsig().  sendsig() does not set cra for
  purecap binaries.

+++ sys/riscv/include/param.h
@@ -40,7 +40,7 @@
 #include <machine/_align.h>
 
 #define	STACKALIGNBYTES	(16 - 1)
-#define	STACKALIGN(p)	((uint64_t)(p) & ~STACKALIGNBYTES)
+#define	STACKALIGN(p)	((uintptr_t)(p) & ~STACKALIGNBYTES)
 
 #ifndef MACHINE
 #define	MACHINE		"riscv"
